refactor: split example project modules
